fix api + started realisation progress bar logic

+++ app/src/main/java/com/android/course/android_course_architecture/domain/model/RecipesModel.kt
@@ -4,6 +4,6 @@ import android.graphics.Bitmap
 
 data class RecipesModel(
     val label: String,
-    val calories: Double,
+    val calories: Int,
     val image: Bitmap?
 )
